requirements.txt
fastapi==0.115.12
starlette==0.46.2
uvicorn==0.34.2
httpx==0.28.1
python-dotenv==1.1.0
pytest==8.3.5
mcp==1.27.0
